requirements.txt
bcrypt==4.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
docopt==0.6.2
Faker==18.7.0
idna==3.4
pipreqs==0.4.13
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
requests==2.30.0
six==1.16.0
urllib3==2.0.2
yarg==0.1.9
datetime==5.1
